requirements.txt
beautifulsoup4==4.12.3
pdfplumber==0.11.4
psycopg[binary]>=3.2.4,<4
python-dotenv==1.0.1
selenium==4.27.1
webdriver-manager==4.0.2
